+++ .eslintrc.yml
@@ -1,0 +1,3 @@
+extends: 'eslint-config-cheminfo'
+env:
+  jest: true
