+++ .travis.yml
@@ -1,0 +1,5 @@
+language: node_js
+node_js:
+  - 8
+  - 6
+
