+++ README.md
@@ -1,9 +1,21 @@
 # logistic-regression
 
+  [![NPM version][npm-image]][npm-url]
+  [![build status][travis-image]][travis-url]
+  [![David deps][david-image]][david-url]
+  [![npm download][download-image]][download-url]
+
 This is an implementation of the logistic regression. When there are more than 2 classes, the method used is the *One VS All*.
 
-## Examples : 
-```javascript
+## Installation
+
+`$ npm install ml-logistic-regression`
+
+## [API Documentation](https://mljs.github.io/logistic-regression/)
+
+## Example
+
+```js
 const {Matrix} = require('ml-matrix');
 
 // our training set (X,Y)
@@ -22,3 +34,17 @@ logreg.train(X,Y);
 var finalResults = logreg.predict(Xtest);
 // Now, you can compare finalResults with the Ytest, which is what you wanted to have.
 ```
+
+
+## License
+
+[MIT](./LICENSE)
+
+[npm-image]: https://img.shields.io/npm/v/ml-logistic-regression.svg?style=flat-square
+[npm-url]: https://npmjs.org/package/ml-logistic-regression
+[travis-image]: https://img.shields.io/travis/mljs/logistic-regression/master.svg?style=flat-square
+[travis-url]: https://travis-ci.org/mljs/logistic-regression
+[david-image]: https://img.shields.io/david/mljs/logistic-regression.svg?style=flat-square
+[david-url]: https://david-dm.org/mljs/logistic-regression
+[download-image]: https://img.shields.io/npm/dm/ml-logistic-regression.svg?style=flat-square
+[download-url]: https://npmjs.org/package/ml-logistic-regression
