+++ src/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const defaultOptions = {};
+
+function logisticRegression(options) {
+    options = Object.assign({}, defaultOptions, options);
+}
+
+module.exports = logisticRegression;
